content_analyzer/requirements.txt
# Core dependencies - minimal set
Flask==2.3.2
pdfplumber==0.10.2
python-docx==0.8.11
Pillow==10.0.1
PyMuPDF==1.23.5
pytesseract==0.3.10
reportlab==4.0.4

# Web server for production
gunicorn==20.1.0

# Minimal Google Cloud dependencies
google-auth==2.22.0
google-api-python-client==2.95.0